Get rid of referential equality, because it's not guaranteed or desirable. Some minor updates to broken serialization unit tests: style, formatting. #45
